+++ env-bootstrap/src/lib.rs
@@ -161,6 +161,7 @@ fn register_panic_hook() {
 fn register_lua_modules() {
     for func in [
         battery::register,
+        color_funcs::register,
         termwiz_funcs::register,
         logging::register,
         mux_lua::register,